+++ client/package.json
@@ -22,7 +22,7 @@
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.6.1",
-    "theme-change": "^2.3.0"
+    "theme-change": "^2.3.0",
     "sweetalert2": "^11.7.0"
   },
   "devDependencies": {
