+++ client/src/components/BtnTheme/index.jsx
@@ -17,35 +17,33 @@ function BtnTheme() {
     }, [])
 
     const themes = [
-        "light",
-        "dark",
-        "cupcake",
-        "bumblebee",
-        "emerald",
-        "corporate",
-        "synthwave",
-        "retro",
-        "cyberpunk",
-        "valentine",
-        "halloween",
-        "garden",
-        "forest",
-        "aqua",
-        "lofi",
-        "pastel",
-        "fantasy",
-        "wireframe",
-        "black",
-        "luxury",
-        "dracula",
-        "cmyk",
-        "autumn",
-        "business",
         "acid",
-        "lemonade",
-        "night",
+        "aqua",
+        "autumn",
+        "black",
+        "bumblebee",
+        "business",
+        "cmyk",
         "coffee",
-        "winter",
+        "corporate",
+        "cupcake",
+        "dark",
+        "dracula",
+        "emerald",
+        "fantasy",
+        "forest",
+        "garden",
+        "halloween",
+        "lemonade",
+        "lofi",
+        "luxury",
+        "night",
+        "pastel",
+        "retro",
+        "synthwave",
+        "valentine",
+        "wireframe",
+        "winter"
     ];
 
     return (
