+++ client/tailwind.config.cjs
@@ -2,34 +2,32 @@
 module.exports = {
   content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
   theme: [
-    "light",
-    "dark",
-    "cupcake",
-    "bumblebee",
-    "emerald",
-    "corporate",
-    "synthwave",
-    "retro",
-    "cyberpunk",
-    "valentine",
-    "halloween",
-    "garden",
-    "forest",
-    "aqua",
-    "lofi",
-    "pastel",
-    "fantasy",
-    "wireframe",
-    "black",
-    "luxury",
-    "dracula",
-    "cmyk",
-    "autumn",
-    "business",
     "acid",
-    "lemonade",
-    "night",
+    "aqua",
+    "autumn",
+    "black",
+    "bumblebee",
+    "business",
+    "cmyk",
     "coffee",
+    "corporate",
+    "cupcake",
+    "dark",
+    "dracula",
+    "emerald",
+    "fantasy",
+    "forest",
+    "garden",
+    "halloween",
+    "lemonade",
+    "lofi",
+    "luxury",
+    "night",
+    "pastel",
+    "retro",
+    "synthwave",
+    "valentine",
+    "wireframe",
     "winter",
   ],
   plugins: [require("daisyui")],
